refactor(computer): move the boot body and the readiness gate (R3, PR-F1b)
`do_boot` and the gate READY is withheld for are the bodies
`Effect::SpawnBoot` and `Effect::SpawnGate` name, so they move next to the
machine that names them. `sandbox::boot::boot_sandbox` still calls both —
PR-F2 is what puts the actor in front of them — so nothing changes about
what runs, in what order, or what it logs.

Verbatim apart from what the new home requires: `super::` is now
`sandbox::`, a few items the body reaches back for widened from
`pub(in crate::sandbox)` to the crate, and the gate's wait became a function
(`wait_for_agent`) instead of a block inside a match arm, with its caller
keeping the failure branch. Proof for the reviewer:

    git show HEAD~1:.../sandbox/boot.rs | sed -n '816,1119p' \
      | sed 's/\bsuper::/sandbox::/g; s/^async fn do_boot(/pub async fn do_boot(/'

diffs empty against the moved range.

The no-await-before-handoff rule is re-stated on `do_boot`'s doc comment
rather than left to the one comment at the site: the signal is what tells a
teardown it may abort the task, and an await between transferring the CoW
handle and sending it is a point where an abort strands a dm device, its
loop device and its COW file with nothing left to name them. Moving code is
exactly when an invariant carried by a single comment gets lost.

Carries PR-E's gap 1: the initial cmd's claim now survives `Gated`. Today
`start_run_workload` takes the workload slot *before* READY — flipping the
instance to Running and publishing RUNNING — and the workload's exit
publishes IDLE. The machine as landed swallowed that claim, so `Gated`
dropped it on the floor: `WorkloadExited` would arrive in `ready` with
nothing to do and the IDLE publish would be lost. `gating` now carries the
reservation, `Gated` lands in `running` when it is taken (arming the idle
window only when the workload exits, which is also what
`arm_idle_timer`'s own `Ready` re-check does today), and a gated computer
projects `Running` to callers — as its instance already does, and as an
Inspect-polling client must see if it is not to steal the slot.

Two tests move with that: the public projection is now a table of allowed
answers per leaf, since `gating` has two, and the stop gate is stated as the
`active` superstate rather than as "whatever projects Ready or Running" —
`gating` projects `Running` while its launch is still in flight, and a stop
there is the actor's `pending_stop`, not the machine's.
