+++ computer/arcbox-computer-runtime/src/lifecycle/tasks.rs
@@ -19,6 +19,8 @@ use std::time::Duration;
 use async_trait::async_trait;
 use tokio::sync::oneshot;
 
+pub mod boot;
+
 use super::effect::ReleaseScope;
 use super::event::{Event, RestoreOrigin};
 use crate::agent::GuestAgent;
